requirements.txt
alembic==1.5.3
aniso8601==8.1.1
astroid==2.4.2
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.0.2
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mccabe==0.6.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-slugify==4.0.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
text-unidecode==1.3
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
